refactor: Delete legacy process instance header
