+++ operate/client/src/App/ProcessInstance/ProcessInstanceHeader/index.tsx
@@ -6,14 +6,10 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {useEffect} from 'react';
 import isNil from 'lodash/isNil';
 import {formatDate} from 'modules/utils/date';
-import {getProcessName} from 'modules/utils/instance';
-import {Operations} from 'modules/components/Operations';
-import {observer} from 'mobx-react';
-import {processInstanceDetailsStore} from 'modules/stores/processInstanceDetails';
-import {variablesStore} from 'modules/stores/variables';
+import {Operations} from 'modules/components/Operations/v2';
+import {getProcessDefinitionName} from 'modules/utils/instance';
 import {Link} from 'modules/components/Link';
 import {Locations, Paths} from 'modules/Routes';
 import {Restricted} from 'modules/components/Restricted';
@@ -23,12 +19,20 @@ import {tracking} from 'modules/tracking';
 import {InstanceHeader} from 'modules/components/InstanceHeader';
 import {Skeleton} from 'modules/components/InstanceHeader/Skeleton';
 import {notificationsStore} from 'modules/stores/notifications';
-import {processStore} from 'modules/stores/process';
 import {VersionTag} from './styled';
 import {useProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
 import {useProcessInstanceXml} from 'modules/queries/processDefinitions/useProcessInstanceXml';
 import {hasCalledProcessInstances} from 'modules/bpmn-js/utils/hasCalledProcessInstances';
-import type {OperationEntityType} from 'modules/types/operate';
+import {type ProcessInstance} from '@vzeta/camunda-api-zod-schemas/8.8';
+import {usePermissions} from 'modules/queries/permissions/usePermissions';
+import {useHasActiveOperations} from 'modules/queries/operations/useHasActiveOperations';
+import {useQueryClient} from '@tanstack/react-query';
+import {PROCESS_INSTANCE_DEPRECATED_QUERY_KEY} from 'modules/queries/processInstance/deprecated/useProcessInstanceDeprecated';
+import {useNavigate} from 'react-router-dom';
+import {useProcessInstanceOperations} from 'modules/hooks/useProcessInstanceOperations';
+import {ProcessInstanceOperationsContext} from './processInstanceOperationsContext';
+import {IS_BATCH_OPERATIONS_V2} from 'modules/feature-flags';
+import {useHasActiveBatchOperationMutation} from 'modules/mutations/processInstance/useHasActiveBatchOperationMutation';
 import {useAvailableTenants} from 'modules/queries/useAvailableTenants';
 
 const headerColumns = [
@@ -77,62 +81,58 @@ const skeletonColumns: {
   },
 ] as const;
 
-const ProcessInstanceHeader: React.FC = observer(() => {
-  const {processInstance} = processInstanceDetailsStore.state;
-  const isMultiTenancyEnabled = window.clientConfig?.multiTenancyEnabled;
-  const processId = processInstance?.processId;
+type Props = {
+  processInstance: ProcessInstance;
+};
+
+const ProcessInstanceHeader: React.FC<Props> = ({processInstance}) => {
+  const queryClient = useQueryClient();
+  const {data: permissions} = usePermissions();
+  const {data: hasActiveOperationLegacy} = useHasActiveOperations();
+  const hasActiveBatchOperationMutation = useHasActiveBatchOperationMutation(
+    processInstance.processInstanceKey,
+  );
   const {
-    state: {process, status},
-  } = processStore;
+    processInstanceKey,
+    processDefinitionVersion,
+    processDefinitionVersionTag,
+    tenantId,
+    startDate,
+    endDate,
+    parentProcessInstanceKey,
+    state,
+    hasIncident,
+    processDefinitionId,
+  } = processInstance;
+  const tenantsById = useAvailableTenants();
+  const tenantName = tenantsById[tenantId] ?? tenantId;
 
-  useEffect(() => {
-    if (processId !== undefined) {
-      processStore.fetchProcess(processId);
-    }
-  }, [processId]);
+  const navigate = useNavigate();
 
-  useEffect(() => {
-    return processStore.reset;
-  }, []);
+  const isMultiTenancyEnabled = window.clientConfig?.multiTenancyEnabled;
 
   const processDefinitionKey = useProcessDefinitionKeyContext();
   const {isPending, data: processInstanceXmlData} = useProcessInstanceXml({
     processDefinitionKey,
   });
+  const {cancellation} = useProcessInstanceOperations(processInstance.state);
 
-  if (
-    processInstance === null ||
-    ['fetching', 'initial'].includes(status) ||
-    isPending
-  ) {
+  if (processInstance === null || isPending) {
     return <Skeleton headerColumns={skeletonColumns} />;
   }
 
-  const {
-    id,
-    processVersion,
-    tenantId,
-    startDate,
-    endDate,
-    parentInstanceId,
-    state,
-    bpmnProcessId,
-  } = processInstance;
-
-  const versionTag = process?.versionTag;
-  const tenantsById = useAvailableTenants();
-  const tenantName = tenantsById[tenantId] ?? tenantId;
-  const versionColumnTitle = `View process "${getProcessName(
+  const versionColumnTitle = `View process "${getProcessDefinitionName(
     processInstance,
-  )} version ${processVersion}" instances${
+  )} version ${processDefinitionVersion}" instances${
     isMultiTenancyEnabled ? ` - ${tenantName}` : ''
   }`;
-  const hasVersionTag = !isNil(versionTag);
+  const hasVersionTag = !isNil(processDefinitionVersionTag);
+  const processInstanceState = hasIncident ? 'INCIDENT' : state;
 
   return (
     <InstanceHeader
-      state={state}
-      hideBottomBorder={state === 'INCIDENT'}
+      state={processInstanceState}
+      hideBottomBorder={hasIncident}
       headerColumns={headerColumns.filter((name) => {
         if (name === 'Tenant') {
           return isMultiTenancyEnabled;
@@ -144,17 +144,17 @@ const ProcessInstanceHeader: React.FC = observer(() => {
       })}
       bodyColumns={[
         {
-          title: getProcessName(processInstance),
-          content: getProcessName(processInstance),
+          title: getProcessDefinitionName(processInstance),
+          content: getProcessDefinitionName(processInstance),
         },
-        {title: id, content: id},
+        {title: processInstanceKey, content: processInstanceKey},
         {
           hideOverflowingContent: false,
           content: (
             <Link
               to={Locations.processes({
-                version: processVersion?.toString(),
-                process: bpmnProcessId,
+                version: processDefinitionVersion?.toString(),
+                process: processDefinitionId,
                 active: true,
                 incidents: true,
                 ...(isMultiTenancyEnabled
@@ -172,7 +172,7 @@ const ProcessInstanceHeader: React.FC = observer(() => {
                 });
               }}
             >
-              {processVersion}
+              {processDefinitionVersion}
             </Link>
           ),
         },
@@ -182,7 +182,7 @@ const ProcessInstanceHeader: React.FC = observer(() => {
                 title: 'User-defined label identifying a definition.',
                 content: (
                   <VersionTag size="sm" type="outline">
-                    {versionTag}
+                    {processDefinitionVersionTag}
                   </VersionTag>
                 ),
               },
@@ -202,20 +202,20 @@ const ProcessInstanceHeader: React.FC = observer(() => {
           dataTestId: 'start-date',
         },
         {
-          title: formatDate(endDate) ?? '--',
-          content: formatDate(endDate),
+          title: formatDate(endDate ?? null) ?? '--',
+          content: formatDate(endDate ?? null),
           dataTestId: 'end-date',
         },
         {
-          title: parentInstanceId ?? 'None',
+          title: parentProcessInstanceKey ?? 'None',
           hideOverflowingContent: false,
           content: (
             <>
-              {parentInstanceId !== null ? (
+              {parentProcessInstanceKey ? (
                 <Link
-                  to={Paths.processInstance(parentInstanceId)}
-                  title={`View parent instance ${parentInstanceId}`}
-                  aria-label={`View parent instance ${parentInstanceId}`}
+                  to={Paths.processInstance(parentProcessInstanceKey)}
+                  title={`View parent instance ${parentProcessInstanceKey}`}
+                  aria-label={`View parent instance ${parentProcessInstanceKey}`}
                   onClick={() => {
                     tracking.track({
                       eventName: 'navigation',
@@ -223,7 +223,7 @@ const ProcessInstanceHeader: React.FC = observer(() => {
                     });
                   }}
                 >
-                  {parentInstanceId}
+                  {parentProcessInstanceKey}
                 </Link>
               ) : (
                 'None'
@@ -240,7 +240,7 @@ const ProcessInstanceHeader: React.FC = observer(() => {
               ) ? (
                 <Link
                   to={Locations.processes({
-                    parentInstanceId: id,
+                    parentInstanceId: processInstanceKey,
                     active: true,
                     incidents: true,
                     canceled: true,
@@ -269,17 +269,29 @@ const ProcessInstanceHeader: React.FC = observer(() => {
         <Restricted
           resourceBasedRestrictions={{
             scopes: ['UPDATE_PROCESS_INSTANCE', 'DELETE_PROCESS_INSTANCE'],
-            permissions: processInstanceDetailsStore.getPermissions(),
+            permissions: permissions,
           }}
         >
-          <>
+          <ProcessInstanceOperationsContext.Provider
+            value={{
+              cancellation: {
+                isPending: cancellation.isPending,
+                onError: () => cancellation.setIsPending(false),
+                onMutate: () => cancellation.setIsPending(true),
+              },
+            }}
+          >
             <Operations
               instance={processInstance}
-              onOperation={(operationType: OperationEntityType) =>
-                processInstanceDetailsStore.activateOperation(operationType)
-              }
-              onError={({operationType, statusCode}) => {
-                processInstanceDetailsStore.deactivateOperation(operationType);
+              onOperation={() => {
+                queryClient.invalidateQueries({
+                  queryKey: [PROCESS_INSTANCE_DEPRECATED_QUERY_KEY],
+                });
+              }}
+              onError={({statusCode}) => {
+                queryClient.invalidateQueries({
+                  queryKey: [PROCESS_INSTANCE_DEPRECATED_QUERY_KEY],
+                });
 
                 notificationsStore.displayNotification({
                   kind: 'error',
@@ -292,26 +304,47 @@ const ProcessInstanceHeader: React.FC = observer(() => {
                 });
               }}
               onSuccess={(operationType) => {
+                queryClient.invalidateQueries({
+                  queryKey: [PROCESS_INSTANCE_DEPRECATED_QUERY_KEY],
+                });
+
                 tracking.track({
                   eventName: 'single-operation',
                   operationType,
                   source: 'instance-header',
                 });
+
+                if (operationType === 'DELETE_PROCESS_INSTANCE') {
+                  navigate(
+                    Locations.processes({
+                      active: true,
+                      incidents: true,
+                    }),
+                    {replace: true},
+                  );
+
+                  notificationsStore.displayNotification({
+                    kind: 'success',
+                    title: 'Instance deleted',
+                    isDismissable: true,
+                  });
+                }
               }}
               forceSpinner={
-                variablesStore.hasActiveOperation ||
-                processInstance?.hasActiveOperation
+                cancellation.isPending ||
+                hasActiveOperationLegacy ||
+                (IS_BATCH_OPERATIONS_V2 && hasActiveBatchOperationMutation)
               }
               isInstanceModificationVisible={
                 !modificationsStore.isModificationModeEnabled
               }
-              permissions={processInstanceDetailsStore.getPermissions()}
+              permissions={permissions}
             />
-          </>
+          </ProcessInstanceOperationsContext.Provider>
         </Restricted>
       }
     />
   );
-});
+};
 
 export {ProcessInstanceHeader};
