+++ operate/client/src/modules/mocks/api/mocks/process.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+const mockProcess = {
+  id: '2251799813688076',
+  name: 'Complex Process',
+  version: 3,
+  bpmnProcessId: 'complexProcess',
+  versionTag: 'myVersionTag',
+};
+
+export {mockProcess};
